pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "patent-mcp-server"
version = "0.1.0"
description = "MCP server for fetching and caching patents by ID"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx[http2]>=0.27",
    "h2>=4.1",
    "tenacity>=8.2",
    "PyMuPDF>=1.24",
    "pdfplumber>=0.11",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "tomli>=2.0; python_version < '3.11'",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
pymupdf4llm = ["pymupdf4llm>=0.0.17"]
browser = ["playwright>=1.44", "playwright-stealth>=1.0"]
marker = ["marker-pdf>=0.3"]
bigquery = ["google-cloud-bigquery>=3.13", "google-auth>=2.23"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "hypothesis>=6.100",
    "reportlab>=4.0",
    "starlette>=0.37",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "anyio>=4.0",
]

[project.scripts]
patent-mcp-server = "patent_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/python/patent_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests/python"]
markers = [
    "browser: requires PATENT_BROWSER_TESTS=1 (Playwright)",
    "integration: hits real network",
    "slow: takes >1s",
]

[tool.hatch.envs.default]
features = ["dev", "pymupdf4llm"]

[tool.hatch.envs.default.scripts]
test = "pytest tests/python -m 'not browser and not integration and not slow' -v"
test-fast = "pytest tests/python -m 'not browser and not integration and not slow' --tb=short -q"
